requirements.txt
python-telegram-bot==21.9
supabase==2.10.0
openai==1.57.4
python-dotenv==1.0.1
openpyxl==3.1.5
pandas==2.2.3
matplotlib==3.9.3
Pillow==11.0.0
httpx==0.28.1
